Implement functions - Make section available and loaded by processor - Add function condition to use functions conditionally

+++ code/lib/src/main/java/org/betonquest/betonquest/lib/function/assignment/StringSourceAssignment.java
@@ -37,7 +37,7 @@ public class StringSourceAssignment implements FunctionAssignment {
 
     @Override
     public boolean asBoolean() {
-        return "true".equalsIgnoreCase(source) || asNumber().intValue() > 0;
+        return "true".equalsIgnoreCase(source) || asNumber().doubleValue() > 0;
     }
 
     @Override
